+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Introduce `OpenEoApiPlainError` for API error responses that are not well-formed
+  for better distinction with properly formed API error responses (`OpenEoApiError`).
+  ([#491](https://github.com/Open-EO/openeo-python-client/issues/491)).
+
 ### Removed
 
 ### Fixed